Added: reducepolicy, reclaimmax, extendpolicy, and stub revokeram
Built out reducepolicy, reclaimmax, and extendpolicy.
Removed time_block for expandpolicy. Use extendpolicy to increase a time_block.
WIP: Put revokeram stub into the .hpp, it is going to be a native action so implementation will
be in the Core.
Renamed reclaimres to reclaimmax, since its reclaiming the max it can from a policy.
Cleaned up some comments and TODOs