Add newuser action to ROA contract
This new action will allow tier-one node owners to create
user accounts for users that are not node owners.

We keep a record of this action for future auditability.

The nonce parameter is intended to help the node owner avoid
accidentally creating more accounts than desired. Currently,
a nonce will never expire once successfully used, although we
may decide to flush the nonce records as part of a network
generation bump.